+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - Add allowNullValues boolean option to pass Null values through helm configs without having them scrubbed (https://github.com/pulumi/pulumi-kubernetes/issues/2089)
+- For SSA conflicts, add a note to the error message about how to resolve (https://github.com/pulumi/pulumi-kubernetes/issues/2235)
 
 ## 3.22.1 (October 26, 2022)