+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add allowNullValues boolean option to pass Null values through helm configs without having them scrubbed (https://github.com/pulumi/pulumi-kubernetes/issues/2089)
 - Fix replacement behavior for immutable fields in SSA mode (https://github.com/pulumi/pulumi-kubernetes/issues/2232)
+- For SSA conflicts, add a note to the error message about how to resolve (https://github.com/pulumi/pulumi-kubernetes/issues/2235)
 
 ## 3.22.1 (October 26, 2022)